+++ types/index.d.ts
@@ -1,3 +1,7 @@
 export * from './string';
 export * from './number';
 export * from './date';
+export * from './validation';
+export * from './array';
+export * from './object';
+export * from './color';