+++ types/validation.d.ts
@@ -1,0 +1,3 @@
+export function isEmail(str: string): boolean;
+export function isURL(str: string): boolean;
+export function isPhoneNumber(str: string): boolean;